Updated some of the install scripts.

+++ scripts/install-plasma.sh
@@ -6,34 +6,34 @@ mkdir -p ~/Applications
 # Prep snap
 #sudo ln -s /var/lib/snapd/snap /snap
 
-toInstall=(
-  firefox
-  gparted
-  solaar
-  qbittorrent
-  vlc
-  pinta
-  cheese
-  meld
-  filezilla
-  inkscape
-  gnome-chess
-  gnome-mines
-  gnome-sudoku
-  aisleriot
-  handbrake
-  audacity
-  blender
-  tiled
-  libreoffice-fresh
-  xorg-xcursorgen
-  plasma-wayland-session
+toInstall="\
+  firefox \
+  gparted \
+  solaar \
+  qbittorrent \
+  vlc \
+  pinta \
+  cheese \
+  meld \
+  filezilla \
+  inkscape \
+  gnome-chess \
+  gnome-mines \
+  gnome-sudoku \
+  aisleriot \
+  handbrake \
+  audacity \
+  blender \
+  tiled \
+  libreoffice-fresh \
+  xorg-xcursorgen \
+  plasma-wayland-session \
   zip
-)
+"
 
 # Arch Linux Setup
 if [ -f "/etc/arch-release" ]; then
-  sudo pacman -S  -q --noconfirm $toInstall
+  sudo pacman -S -q --noconfirm $toInstall
 
   # VS Code (Arch)
   git clone https://aur.archlinux.org/visual-studio-code-bin
@@ -69,29 +69,36 @@ flatpak install flathub \
     --assumeyes
 
 # pCloud
-wget https://p-def8.pcloud.com/cBZu1eKrMZMqwjTqZZZh9ILr7Z2ZZmNzZkZPF7pVZrHZrzZBpZpRZc5Z6pZopZ9pZApZGFZUJZkJZPpZkzZWTVkVZQDtD9nHvjKu7c1tTLeXYVj05EfNk/pcloud -O ~/Applications/pcloud
-chmod +x ~/Applications/pcloud
-~/Applications/pcloud &
+if [ ! -f "~/Applications/pcloud" ]; then
+  wget https://p-def8.pcloud.com/cBZu1eKrMZMqwjTqZZZh9ILr7Z2ZZmNzZkZPF7pVZrHZrzZBpZpRZc5Z6pZopZ9pZApZGFZUJZkJZPpZkzZWTVkVZQDtD9nHvjKu7c1tTLeXYVj05EfNk/pcloud -O ~/Applications/pcloud
+  chmod +x ~/Applications/pcloud
+  ~/Applications/pcloud &
+fi
 
 # PIA
-wget https://installers.privateinternetaccess.com/download/pia-linux-3.3.1-06924.run -O ~/Downloads/pia.run
-chmod +x ~/Downloads/pia.run
-mkdir -p ~/.config/privateinternetaccess
-cp ./scripts/pia-settings.json ~/.config/privateinternetaccess/clientsettings.json
-~/Downloads/pia.run
-rm ~/Downloads/pia.run
+if ! command -v piactl &> /dev/null
+then
+  wget https://installers.privateinternetaccess.com/download/pia-linux-3.3.1-06924.run -O ~/Downloads/pia.run
+  chmod +x ~/Downloads/pia.run
+  mkdir -p ~/.config/privateinternetaccess
+  cp ./scripts/pia-settings.json ~/.config/privateinternetaccess/clientsettings.json
+  ~/Downloads/pia.run
+  rm ~/Downloads/pia.run
+fi
 
 # Moe Theme
 wget https://gitlab.com/jomada/moe-theme/-/archive/master/moe-theme-master.tar.gz -O moe.tar.gz
 tar -xzvf ./moe.tar.gz
 
 # Dark
+mkdir -p ~/.local/share/color-schemes/MoeDark.colors
 cp -r ./moe-theme-master/MoeDark-Global/* ~/.local/share/plasma/look-and-feel/
 cp -r ./moe-theme-master/MoeDark ~/.local/share/plasma/desktoptheme/MoeDark
 cp -r ./moe-theme-master/MoeDark-aurorae/* ~/.local/share/aurorae/themes/
 cp ./moe-theme-master/Moe-Dark-color-schemes/MoeDark.colors ~/.local/share/color-schemes/MoeDark.colors
 
 # Light
+mkdir -p ~/.local/share/color-schemes/Moe.colors
 cp -r ./moe-theme-master/Moe-Global/* ~/.local/share/plasma/look-and-feel/
 cp -r ./moe-theme-master/Moe ~/.local/share/plasma/desktoptheme/Moe
 cp -r ./moe-theme-master/aurorae/* ~/.local/share/aurorae/themes/
@@ -101,17 +108,21 @@ cp ./moe-theme-master/color-schemes/Moe.colors ~/.local/share/color-schemes/Moe.
 lookandfeeltool -a Moe-Dark
 
 # Iosevka Font
+mkdir -p ~/.fonts
 wget "https://objects.githubusercontent.com/github-production-release-asset-2e65be/39315600/97e975fb-b154-4b05-bd0a-3c818fdb02f9?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=AKIAIWNJYAX4CSVEH53A%2F20220424%2Fus-east-1%2Fs3%2Faws4_request&X-Amz-Date=20220424T063917Z&X-Amz-Expires=300&X-Amz-Signature=4edc659fe314a1f8b8964063152d98f2ce1e1c971619d87c08996443ce65cf68&X-Amz-SignedHeaders=host&actor_id=0&key_id=0&repo_id=39315600&response-content-disposition=attachment%3B%20filename%3Dsuper-ttc-iosevka-ss09-15.2.0.zip&response-content-type=application%2Foctet-stream" -O ./iosevka-ss09.zip
 bsdtar xvf ./iosevka-ss09.zip
 mkdir -p ~/.fonts
 cp -r ./*.ttc ~/.fonts/
 
 # Papirus Icons
+mkdir -p ~/.local/share/icons
 wget -qO- https://git.io/papirus-icon-theme-install | DESTDIR="$HOME/.local/share/icons" sh
 /usr/lib/plasma-changeicons Papirus-Dark
 
 # simp1e cursors
-git clone --recurse-submodules https://gitlab.com/zoli111/simp1e.git
+if [ ! -d "simp1e" ] ; then
+  git clone --recurse-submodules https://gitlab.com/zoli111/simp1e.git
+fi
 cd simp1e
 find . -type f -name "*.sh" -exec sed -i 's/\r//' {} \;
 chmod +x ./generate_svgs.sh
@@ -134,7 +145,9 @@ for (i=0;i<Desktops.length;i++) {
 }"
 
 # Kickstart / Start Menu, installed but not configured
-git clone https://gitlab.com/nwwdles/plasma-kickoff-grid.git
+if [ ! -d "plasma-kickoff-grid" ] ; then
+  git clone https://gitlab.com/nwwdles/plasma-kickoff-grid.git
+fi
 cd plasma-kickoff-grid
 zip -r "plasma-kickoff-grid.plasmoid" package
 kpackagetool5 --install "plasma-kickoff-grid.plasmoid"
